fixed multiple file errors
