--- public/README.md
@@ -1,13 +1,0 @@
-# TodoList-EI
-This is a ToDo List application using React.
-
-## Instructions to run the app
-1. Clone this repository: `git clone https://github.com/fosa-ij/TodoList-EI
-2. cd TodoList-EI
-3. npm install
-4. npm run dev
-
-## To contribute to this repository: 
-- Create a new branch 
-- Add files changed and commit to that branch 
-- Push your fixed code to that branch and create a pull request to development
