+++ src/Todo.css
@@ -1,9 +1,139 @@
+div .todo-container {
+    box-sizing: border-box;
+    width: 100%;
+    margin-bottom: 20px;
+    padding: 10px 20px;
+    border-radius: 10px;
+    background-color: #fff;
+    box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
+}
+
+.todo-container hr {
+    opacity: 0.3;
+    margin-bottom: 20px;
+}
+
 .display-flex {
     display: flex;
     justify-content: space-between;
     align-items: center;
+    /* margin-bottom: -7px; */
 }
 
-.todo-container {
-    width: 25%;
+.todo-main {
+    text-align: left;
+}
+
+.todo-main p {
+    font-weight: 600;
+    margin-bottom: 3px;
+}
+
+.todo-main > .tag {
+    font-size: 11px;
+    padding: 4px 10px;
+    border-radius: 15px;
+    background-color: #ffdbdb;
+}
+
+
+/*=============================
+ Checkbox Styling
+ ============================== */
+
+.check > [type=checkbox] {
+    width: 2rem;
+    height: 2rem;
+    color: dodgerblue;
+    vertical-align: middle;
+    -webkit-appearance: none;
+    background: none;
+    border: 0;
+    outline: 0;
+    flex-grow: 0;
+    border-radius: 50%;
+    background-color: #FFFFFF;
+    transition: background 300ms;
+    cursor: pointer;
+  }
+
+  .check > [type=checkbox]::before {
+    content: "";
+    color: transparent;
+    display: block;
+    width: inherit;
+    height: inherit;
+    border-radius: inherit;
+    border: 0;
+    background-color: transparent;
+    background-size: contain;
+    box-shadow: inset 0 0 0 1px #CCD3D8;
+  }
+  
+ 
+  .check > [type=checkbox]:checked {
+    background-color: currentcolor;
+  }
+  
+  .check > [type=checkbox]:checked::before {
+    box-shadow: none;
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24'%3E %3Cpath d='M15.88 8.29L10 14.17l-1.88-1.88a.996.996 0 1 0-1.41 1.41l2.59 2.59c.39.39 1.02.39 1.41 0L17.3 9.7a.996.996 0 0 0 0-1.41c-.39-.39-1.03-.39-1.42 0z' fill='%23fff'/%3E %3C/svg%3E");
+  }
+  
+/* IE */
+
+[type=checkbox]::-ms-check {
+    content: "";
+    color: transparent;
+    display: block;
+    width: inherit;
+    height: inherit;
+    border-radius: inherit;
+    border: 0;
+    background-color: transparent;
+    background-size: contain;
+    box-shadow: inset 0 0 0 1px #CCD3D8;
+  }
+  
+  [type=checkbox]:checked::-ms-check {
+    box-shadow: none;
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24'%3E %3Cpath d='M15.88 8.29L10 14.17l-1.88-1.88a.996.996 0 1 0-1.41 1.41l2.59 2.59c.39.39 1.02.39 1.41 0L17.3 9.7a.996.996 0 0 0 0-1.41c-.39-.39-1.03-.39-1.42 0z' fill='%23fff'/%3E %3C/svg%3E");
+  }
+
+  /* Date */
+
+.foot {
+    padding-bottom: 10px;
+}
+
+.date {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.date span {
+    font-size: 13px;
+    font-weight: 700;
+    opacity: 0.4;
+}
+
+/* Action Buttons */
+
+.btn-atn {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.btn-atn button {
+    background-color: transparent;
+    border: none;
+    font-size: 18px;
+    opacity: 0.5;
+}
+
+.btn-atn .del-btn {
+    opacity: 1;
+    color: #f62d2d;
 }
