--- src/assets/randomNumbers.js
@@ -1,22 +1,0 @@
-let haveIt = [];
-
-export default function generateUniqueRandom(maxNr) {
-    //Generate random number
-    let random = (Math.random() * maxNr).toFixed();
-
-    //Coerce to number by boxing
-    random = Number(random);
-
-    if(!haveIt.includes(random)) {
-        haveIt.push(random);
-        return random;
-    } else {
-        if(haveIt.length < maxNr) {
-          //Recursively generate number
-         return  generateUniqueRandom(maxNr);
-        } else {
-          console.log('No more numbers available.')
-          return false;
-        }
-    }
-}
