+++ src/index.css
@@ -1,9 +1,141 @@
-:root {
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+/* family=Montserrat :wght@300;400;600;700  */
 
+/*===========================================
+Resources://
+Date Format: https://stackoverflow.com/questions/34590369/formatting-a-date-string-in-react-native
+Add Google Fonts: https://stackoverflow.com/questions/40769551/how-to-use-google-fonts-in-react-js
+Gradient: https://cssgradient.io/gradient-backgrounds/
+Shadow: https://getcssscan.com/css-box-shadow-examples
+============================================*/
+
+:root {
+  font-family: 'Montserrat', system-ui, Avenir, Helvetica, Arial, sans-serif;
   font-synthesis: none;
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   -webkit-text-size-adjust: 100%;
+  /* background-color: #f1f2f6; */
+  background-color: #8EC5FC;
+  background-image: linear-gradient(62deg, #8EC5FC 0%, #E0C3FC 100%);
+  color: #2f3542;
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+/* ====================================================
+Styling for <Input /> Headings and Time 
+=======================================================*/
+
+.main-container {
+  background-color: #f1f2f6;
+  max-width: 40%;
+  margin: 0 auto;
+  padding: 10px 20px;
+  text-align: center;
+  border-radius: 10px;
+}
+
+.main-container .container-head {
+  margin-bottom: 30px;
+}
+
+.main-container .container-head h1 {
+  font-size: 30px;
+  font-weight: 600;
+  margin-bottom: -10px;
+}
+
+.main-container .container-head p {
+  font-weight: 600;
+  color: #a0a0a0;
+}
+
+/* ====================================================
+Styling for <Input />  
+=======================================================*/
+
+section .todo-add {
+  width: 100%;
+}
+
+input[type=text]  {
+    display: block;
+    width: 100%;
+    padding: 15px 10px;
+    margin-bottom: 15px;
+    box-sizing: border-box;
+    border: 1px solid #0560f284;
+    border-radius: 8px;
+    background-color: #ffffff;
+    transition: 0.3s ease-in-out;
+    
+} 
+
+input[type=text]:focus {
+  outline: none;
+  background-color: #cee2ff;
+  border: 1px solid #055ef2;
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 12px;
+}
+
+.tags {
+  display: flex;
+  align-items: center;
+  justify-content: left;
+  width: 100%;
+  gap: 10px;
+  margin-bottom: 30px;
+  
+}
+
+.urgent, .important {
+  padding: 3px 20px 4px 20px; 
+  border-radius: 30px;
+  background-color: #d4d4d4;
+  font-size: 12px;
+  font-weight: 400;
+  color: #000; 
+}
+
+/* Add Todo Button */
+.todo-add > button {
+  width: 50%;
+  padding: 13px 0px;
+  border: none;
+  border-radius: 8px;
+  background-color: #055ef2;
+  color: #fff;
+  font-weight: 600;
+  font-size: 13px;
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 12px;
+  cursor: pointer;
+}
+
+/* Divider */
+
+.divider {
+  margin-top: 30px;
+}
+
+.divider hr {
+  width: 50%;
+  opacity: 0.3;
+}
+
+/* ================================
+Responsive styles 
+================================= */
+/* Tablet */
+
+@media only screen and (max-width: 890px) {
+  .main-container {
+    max-width: 70%;
+  }
+}
+
+@media only screen and (max-width: 430px) {
+  .main-container {
+    max-width: 80%;
+  }
 }
